swap leader and civ classes
