+++ server/src/game/civilization.ts
@@ -1,77 +1,85 @@
-import { Leader, LeaderData } from './leader';
-import { Unit } from './map/tile/unit';
-import { Coords } from './world';
+import { KnowledgeMap } from './map/tile/knowledge';
+
+const DEFAULT_START_KNOWLEDGE = {
+  'start': 100,
+  'food_0': 100,
+  'military_0': 100,
+  'science_1': 100,
+}
+
+export const civTemplates = [
+  { color: '#820000', textColor: '#ccc', name: 'Rokun', startingKnowledge: { 'science_1': 0, 'military_1': 100 } }, // RICH RED
+  { color: '#0a2ead', textColor: '#ccc', name: 'Azura' }, // BLUE
+  { color: '#03a300', textColor: '#222', name: 'Vertos' }, // GREEN
+  { color: '#bd9a02', textColor: '#222', name: 'Solei' }, // SAND YELLOW
+  { color: '#560e8a', textColor: '#ccc', name: 'Imperius' }, // ROYAL PURPLE
+  { color: '#bd7400', textColor: '#333', name: 'Baranog' }, // ORANGE
+];
 
 export interface CivilizationData {
-  color?: string;
-  leader?: LeaderData;
+  id: number;
+  color: string;
+  textColor: string;
+  secondaryColor: string;
+  name: string;
+  leaderID: number | null;
 }
 
 export class Civilization {
-  units: Unit[];
-  leader?: Leader;
-  turnActive: boolean;
-  turnFinished: boolean;
+  private id: number;
+  private color: string;
+  private textColor: string;
+  private secondaryColor: string;
+  private name: string;
+  private leaderID: number | null;
 
-  constructor() {
-    this.units = [];
-    this.turnActive = false;
-    this.turnFinished = false;
-  }
+  public startingKnowledge: KnowledgeMap;
 
-  export() {
-    return {
-      units: this.units.map(unit => unit.export()),
-      turnActive: this.turnActive,
-      turnFinished: this.turnFinished,
+  constructor(id: number) {
+    const { color, textColor, name, startingKnowledge } = civTemplates[id];
+    this.id = id;
+    this.color = color;
+    this.textColor = textColor;
+    this.secondaryColor = color;
+    this.name = name;
+    this.leaderID = null;
+
+    this.startingKnowledge = {
+      ...DEFAULT_START_KNOWLEDGE,
+      ...startingKnowledge,
     };
   }
 
   static import(data: any): Civilization {
-    const civ =  new Civilization();
-    // civ.units = data.units.map(unitData => Unit.import(unitData));
-    civ.turnActive = data.turnActive;
-    civ.turnFinished = data.turnFinished;
+    const civ = new Civilization(data.id);
+    civ.color = data.color;
+    civ.textColor = data.textColor;
+    civ.secondaryColor = data.secondaryColor;
+    civ.name = data.name;
+    civ.leaderID = null;
     return civ;
   }
 
+  select(leaderID: number): void {
+    this.leaderID = leaderID;
+  }
+
+  unselect(): void {
+    this.leaderID = null;
+  }
+
+  hasLeader(): boolean {
+    return this.leaderID !== null;
+  }
+
   getData(): CivilizationData {
-    const leaderData = this.leader?.getData();
     return {
-      color: leaderData?.color,
-      leader: leaderData,
-    }
-  }
-
-  newTurn() {
-    this.turnActive = true;
-    this.turnFinished = false;
-
-    for (const unit of this.units) {
-      unit.newTurn();
-    }
-  }
-
-  endTurn() {
-    this.turnActive = false;
-  }
-
-  getUnits(): Unit[] {
-    return this.units;
-  }
-
-  getUnitPositions(): Coords[] {
-    return this.units.map(unit => unit.coords);
-  }
-
-  addUnit(unit: Unit): void {
-    this.units.push(unit);
-  }
-
-  removeUnit(unit: Unit): void {
-    const unitIndex = this.units.indexOf(unit);
-    if (unitIndex > -1) {
-      this.units.splice(unitIndex, 1);
-    }
+      id: this.id,
+      color: this.color,
+      textColor: this.textColor,
+      secondaryColor: this.secondaryColor,
+      name: this.name,
+      leaderID: this.leaderID,
+    };
   }
 }
